Basic multi-user support.
Every lesson now has to belong to a user. In every moment a user
has to be signed in and he sees only his own lessons. No security
measures are applied though. User log in using a unique UUID token
by providing it as a login_token argument to the application URL.
The user name is displayed right below the application name. Please
note that after signing in, you have to refresh the page to see the
displayed user name change. If no user is signed in, the demo user
is signed in automatically.