Function to split a table into many inherited ones.
The parent table remains in the schema, but it is meant to be
empty. All the data are going to be stored in the children tables.
Thanks to the table inheritance, the records can be manipulated
simply by referencing the parent table, only inserts have to be
done to the specific children table.

Also introduces a new DbObject class which stores the non-
-escaped name of the table/column name, but when treated
as a string, the escaped name is provided. This allows all the
functions to simply use the name in the queries, but if they’d
need to work with the raw name, it is retrievable.