Only superusers can manage languages.
A superuser flag added to the users table. Superusers can edit and
delete languages as well as mark them as default. Non-super users
can only add new languages, but they are not allowed to manage
them. This applies even to the languages that non-superusers have
added themselves.

Glutexo is a superuser, demo is not. This should prevent vandalism.